+++ packages/react-ui/src/components/atoms/LabelWithIndicator.js
@@ -5,6 +5,7 @@ import { styled } from '@mui/material/styles';
 import Typography from './Typography';
 
 const LabelIndicator = styled(Typography)(({ theme }) => ({
+  marginLeft: theme.spacing(0.5),
   '.Mui-disabled &': {
     color: theme.palette.text.disabled
   }